Catch empty db on randomkey

+++ src/page_index.h
@@ -31,7 +31,7 @@ int redislite_delete_key(void *_cs, void *first_page, char *key, size_t length, 
 int redislite_delete_keys(void *_cs, int q, char **keys, size_t *lengths);
 int redislite_exists_key(void *_db, void *_cs, void *first_page, char *key, size_t length);
 int redislite_get_keys(void *_db, void *_cs, char *pattern, int pattern_len, int *number_of_keys_p, char ***keys_p, int **keys_length_p);
-int redislite_get_random_key_name(void *_db, void *_cs, char **key_p, int *key_length_p);
+int redislite_get_random_key_name(void *_db, void *_cs, char **key_p, size_t *key_length_p);
 int redislite_page_index_rename_key(void *_cs, void *first_page, char *src, size_t src_len, char *target, size_t target_len);
 int redislite_page_index_renamenx_key(void *_cs, void *first_page, char *src, size_t src_len, char *target, size_t target_len);
 int redislite_flush(void *_cs);
